Use new Boots API of JA
In December of 2020, I added a Boots API method to the JsonAssets API specifically to support STF.

Use the new Boots API method of the JA API, if it is available.  If not, fall back to the non-Boots version of the API.  The effect is to be able to add all boots of a JA pack when the JA API supports it.