API changes to support JsonAssets boots and CustomFurniture (#44)
* CP support for Custom Furniture

Allow items in content packages for Custom Furniture to be added or removed via the JAPacks fields.  This will be harmless (except for the usual message when an API can't be reached) until the corresponding pull request is approved for Custom Furniture to actually provide this API.

The effect of this change is to 1) allow ItemStocks to be initialized using all the contents of a Custom Furniture content pack mentioned in the JAPacks field, and 2) allow all the contents of a Custom Furniture content pack to be removed from the usual in-game shop (Robin's carpentry shop, in this case).

I should also note that the initial null checks for the JsonAssets API were removed from the pack-handling methods; this is because these methods can now support packs of both JA and CF.  Instead, suitable protection against null APIs is provided in these methods via the ?. operator.

At this point, CF furniture does not yet show up in shops due to them not yet being available via ItemsUtil.UpdateObjectInfoSource.  Getting that working likely requires further changes to Custom Furniture itself so that the furniture content is available when UpdateObjectInfoSource is invoked.

Also, made a change to GetApi so that the same instance of the STF API is always returned.  The STF API is still a work in progress, but it should be a little better now.

* Use new Boots API of JA

In December of 2020, I added a Boots API method to the JsonAssets API specifically to support STF.

Use the new Boots API method of the JA API, if it is available.  If not, fall back to the non-Boots version of the API.  The effect is to be able to add all boots of a JA pack when the JA API supports it.

Co-authored-by: Yiminxue Zheng <33401364+ChroniclerCherry@users.noreply.github.com>